wxOS2 night build warning fixes.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@38108 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/os2/listbox.cpp
@@ -608,7 +608,7 @@ void wxListBox::SetString(int N, const wxString& rsString)
 
     int                             nNewN = N;
 
-    if (N == m_nNumItems - 1)
+    if (N == (int)(m_nNumItems - 1))
         nNewN = -1;
 
     ::WinSendMsg( GetHwnd()
